edit file migrate: status category + product_rate(int)

+++ database/migrations/2018_10_29_071539_create_rates_table.php
@@ -19,7 +19,7 @@ class CreateRatesTable extends Migration
             $table->unsignedInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
             $table->foreign('product_id')->references('id')->on('products');
-            $table->integer('rate_mark');
+            $table->float('rate_mark',8,2);
         });
     }
